loading levels and game completion
added functions that will choose to load specific sql queries that willl load the proper data from the proper table. Created a genreic functino to populate the table with data taken from php. added booleans to tell if 13x13 or 7x7 is active and added that to check victory function. Added the ability to alert the screen if we won with the time until completion.